Fixed Bug #{4600}
Done.

The file format is called Neighborhood Polygon File and is home brewed.

A sample .npf file is included in the distribution in the polygons subdirectory of the installation.

The "Test Data" button is replaced with a "Browse" button on the wizard that will allow you to browse the local disk for .npf file.

Some things to keep in mind when testing:
* There is little to no validation being done on the .npf file since this is a stop gap and will (hopefully) be replaced with the more robust World Database solution.
* You can edit the .npf and change the names of polygons, remove children and put them back, but be careful.
* Polygons can now have the same display name, IDs are used internally to keep them straight.

First, have a look at SE_Asia.npf in the polygons directory and take note of the fields in a "poly" record.
Here's what a record looks like and what the fields mean:

poly {
    id       N
    polyset  filename
    setnum   M
    name     string
    ptype    string
    children list of strings
    parent   string
}

Where 
   id       - the nth polygon to read, this ID must correspond to the
              position of the polygon data in the KML file
   polyset  - the name of the KML file to read this polygons data from 
   setnum   - an integer ID for the KML file, not yet used
   name     - the name to display in the wizard for the nth polygon read
   ptype    - free form string for polygon type (ie. National, District)
   children - list of names of other polygons

To remove a child you must remove the "poly" definition of it and the reference to it as a child in the parent.
To add it back undo that.
To change the name displayed in the GUI change the "name" field.

Note that "id" is fragile and is how the metadata in an .npf is properly associated with polygon data in its corresponding KML file. You don't want to mess with "id".

Still to do:
  - better validation of an .npf (dunno how far this will go)
  - a better way to associate polygon coordinate data to polygon metadata