Fixed Bug #{4598}
Rather than fix the popup dialog I've decided to go a different route which still fixes the problem but makes the GUI nicer.

The top of the Neighborhood Ingestion wizard now shows some information about the playbox and the number of polygons read and the number in the playbox.

Retest.