Fixed Bug #{4564}
Done.  The "absit.*" parms were out of position in both the object and 
the help trees in the detail browser's side bar.

To fix the Help tree, I repositioned the absit.* parameters in the 
parmdb.tcl, since that tree is built according to the order of definition.
To fix the Object tree, I edited the list of parm sets in appserver_parmdb.tcl.

Both are now fine.