Fixed Bug #{4604}
This bug is a result of a change to how the method that returns selected polygons figures out what to extract.

Essentially, even though the all the contents of the tree are hidden because no neighborhoods are in the playbox, the top level neighborhoods are incorrectly output.

A check is now in that method to make sure it no longer outputs neighborhoods that are not visible.